src/requirements.txt
openai>=1.3.0
pandas>=1.5.0
python-dotenv>=1.0.0
numpy>=1.21.0
scikit-learn>=1.3.0
matplotlib>=3.5.0
kneed>=0.8.0
hdbscan>=0.8.0
umap-learn>=0.5.0
python-igraph>=0.10.0
leidenalg>=0.9.0
scipy>=1.7.0
leidenalg>=0.9.0